address comments on refactor
