+++ python/mxnet/ndarray/_internal.py
@@ -16,4 +16,22 @@
 # under the License.
 
 """NDArray namespace used to register internal functions."""
-__all__ = []
+import sys as _sys
+import os as _os
+try:
+    if int(_os.environ.get("MXNET_ENABLE_CYTHON", True)) == 0:
+        from .._ctypes.ndarray import NDArrayBase, CachedOp
+        from .._ctypes.ndarray import _set_ndarray_class, _imperative_invoke
+    elif _sys.version_info >= (3, 0):
+        from .._cy3.ndarray import NDArrayBase, CachedOp
+        from .._cy3.ndarray import _set_ndarray_class, _imperative_invoke
+    else:
+        from .._cy2.ndarray import NDArrayBase, CachedOp
+        from .._cy2.ndarray import _set_ndarray_class, _imperative_invoke
+except ImportError:
+    if int(_os.environ.get("MXNET_ENFORCE_CYTHON", False)) != 0:
+        raise ImportError("Cython Module cannot be loaded but MXNET_ENFORCE_CYTHON=1")
+    from .._ctypes.ndarray import NDArrayBase, CachedOp
+    from .._ctypes.ndarray import _set_ndarray_class, _imperative_invoke
+
+__all__ = ['NDArrayBase', 'CachedOp', '_imperative_invoke', '_set_ndarray_class']
