+++ python/mxnet/symbol/op.py
@@ -19,18 +19,3 @@
 # pylint: disable=unused-import
 """Backend ops in mxnet.symbol namespace."""
 __all__ = []
-
-import sys as _sys
-import os as _os
-
-try:
-    if int(_os.environ.get("MXNET_ENABLE_CYTHON", True)) == 0:
-        from .._ctypes.symbol import SymbolBase, _set_symbol_class
-    elif _sys.version_info >= (3, 0):
-        from .._cy3.symbol import SymbolBase, _set_symbol_class
-    else:
-        from .._cy2.symbol import SymbolBase, _set_symbol_class
-except ImportError:
-    if int(_os.environ.get("MXNET_ENFORCE_CYTHON", False)) != 0:
-        raise ImportError("Cython Module cannot be loaded but MXNET_ENFORCE_CYTHON=1")
-    from .._ctypes.symbol import SymbolBase, _set_symbol_class
